requirements.txt
jsonrpcclient[requests]
requests
discord.py
python-dotenv